xds: distribute notifications more fairly
This decreases the chance that one slow client will block other clients.  An alterantive is to
buffer the session channel, but that really just pushes the problem out until "later".  If a buffer
of size 0 doesn't work, a buffer of size 1 won't work -- it will just work more frequently until
sometime that fails too.  (It also hides concurrency bugs around the fact that sendUpdate() and
xds.Add/Update/Delete() both compete for the same Manager lock.)