xds: clean up resource/session locking a little bit
The previous changes made blocking even worse, but I think the program more correct, so that's good.
This change addresses a few things.

1) notify doesn't care if the resources change while it's notifying.  It only cares that a stream
doesn't close its channel while it's trying to notify it.  So we split the locking into resource
locks and session locks, so that as soon as a stream wakes up because of a resource changes, it can
read the resources while notify is sending.  (The functions that call notify obviously lock the
resources correctly, of course.)

2) We are more paranoid about the situation where a stream exits, but it's not reading from its
session because it's waiting for the session lock, where it can then close the session.  We just
start a goroutine to drain the session while waiting for the lock, the goroutine then exits when we
take the lock and close the channel.  I think this guarantees correctness (we should never get
"panic: send on closed channel", not that we've ever seen that in ekglue, because we only send to
channels and close channels when we have the mutex.)

There must be a better design, though.  Mixing mutexes and channels feels real bad.